+++ CHANGELOG.md
@@ -76,6 +76,7 @@ This new version of the app is a rewrite of Learn GDScript in Godot 4 that refin
 - Use debug templates in export_presets & defer run to prevent crashes of verification function (#1222)
 - Fix matching brace close not working correctly on macOS desktop (#712)
 - Fix practice scene not loading on Chrome browser on macOS (#583)
+- Fix placement issues in particular for long glossary entries that could end up cropped (#932)
 - L7.P2: Fix turtle position not resetting after running code (#866)
 - L20.P1: Fix practice validation logic (#1212)
 - Fix crash in L23 when running certain code caused the program to close (#826)